# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Watch Store API List",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/send_sms": {
            "post": {
                "tags": [
                    "Auth Api"
                ],
                "description": "use for send verification sms to user",
                "operationId": "34907421430decd0ccbdc9877b69a346",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "mobile": {
                                        "description": "Enter mobile number",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/verify_sms_code": {
            "post": {
                "tags": [
                    "Auth Api"
                ],
                "description": "use to check sms code that recieved by user",
                "operationId": "e1675d3d6afe52cfa674dfe6374b6239",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "mobile": {
                                        "description": "user mobile number",
                                        "type": "string"
                                    },
                                    "code": {
                                        "description": "recieved user sms code",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Bearer",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}